Migrate the TopBanner to tailwind
