+++ src/components/tailwind/index.js
@@ -14,6 +14,7 @@ export { default as TextLink } from './TextLink';
 export { default as ResponsiveSpacer } from './ResponsiveSpacer';
 export { default as InterstitialTitle } from './InterstitialTitle';
 export { default as CodeBlock } from './CodeBlock';
+export { default as TopBanner } from './TopBanner';
 export {
   DotPattern,
   SplitDiagonalPattern,
